Add menu and battle scenes
