+++ src/phaser/scenes/battleScene.js
@@ -1,0 +1,11 @@
+import Phaser from 'phaser';
+
+export default class BattleScene extends Phaser.Scene {
+  constructor() {
+    super({ key: 'battle' });
+  }
+
+  create() {
+    this.cameras.main.setBackgroundColor('#2e2e2e');
+  }
+}
